Add processing of requests to edit and delete posts and comments; start working with tags.

+++ server/src/main/java/com/example/server/dto/ContentObjDto.java
@@ -1,15 +1,15 @@
 package com.example.server.dto;
 
 public class ContentObjDto {
-    private final long authorId;
+    private final int authorId;
     private final String content;
 
-    public ContentObjDto(long authorId, String content) {
+    public ContentObjDto(int authorId, String content) {
         this.authorId = authorId;
         this.content = content;
     }
 
-    public long getAuthorId() {
+    public int getAuthorId() {
         return authorId;
     }
